ims_usrloc_pcscf: rename gruu history table; GPL; drop UNIT_TEST stubs
Use pcscf_gruu_history in DB code. Add SPDX GPL headers to new sources.
Remove unit-test ifdefs from production files; provide stub headers under
test/modules for standalone unit builds.
