+++ src/modules/ims_usrloc_pcscf/impu_match.c
@@ -1,19 +1,31 @@
-/* impu_match.c */
+/*
+ * Copyright (C) 2026 toharishs@gmail.com
+ *
+ * The initial version of this code is written by Harish S
+ *
+ * SPDX-License-Identifier: GPL-2.0-or-later
+ *
+ * Kamailio is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version
+ *
+ * Kamailio is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ *
+ */
+
+/* Matches an AOR/IMPU pair, including SIP/tel user-part comparison */
+
 #include "impu_match.h"
 #include <string.h>
-
-#ifdef UNIT_TEST
-/* minimal sip_uri declaration for unit tests to avoid heavy parser includes */
-struct sip_uri
-{
-	int type;
-	str user;
-	str host;
-};
-int parse_uri(char *buf, int len, struct sip_uri *uri);
-#else
 #include "../../core/parser/parse_uri.h"
-#endif
 
 int pcscf_impu_matches_aor(str *aor, str *impu)
 {
