+++ test/modules/ims_usrloc_pcscf/test_temp_gruu_lru.c
@@ -1,5 +1,6 @@
 #include "test_common.h"
 #include "pcontact_index.h"
+#include "usrloc.h"
 #include <time.h>
 
 static int test_lru_add_get(void)
